Rename Services and Products Controllers
